checkpoint
refactoring, trying to remove camPos in draw functions, should be able to translate via matrix

+++ gameEngine/gameEngine/SensorActuators/MotionActuator.h
@@ -5,18 +5,18 @@
 class MotionActuator : public AbstractActuator
 {
 public:
-	MotionActuator(std::string name, MovingSceneObject*, int newDX, int newDY);
+	MotionActuator(std::string name, MovingSceneObject*, float newDX, float newDY);
 	MotionActuator(std::string name, MovingSceneObject*, double newAngle, double newForce, std::string condition = "force");
-	MotionActuator(std::string name, MovingSceneObject*, int newDX, std::string condition);
+	MotionActuator(std::string name, MovingSceneObject*, float newDX, std::string condition);
 	MotionActuator(std::string name, MovingSceneObject*, std::string condition);
 	void run();
 	~MotionActuator();
 
 private:
 	MovingSceneObject *object;
-	int DX;
-	int DY;
-	int DT;
+	float DX;
+	float DY;
+	float DT;
 	double force;
 	double angle;
 	std::string condition;
